Refactor script, be careful to restore defaults
This commit refactors portions of the code to make it behave better;
i.e. no defaults are changed before the user has the chance to select
the inner and outer edges. This way, if the selection fails, no defaults
are left unchanged. This commit also saves and restores the value of the
[$domain getUnstructuredSolverAttribute ShapeConstraint] option that is
assigned to the created domain. This option is disabled (set to free)
during the execution of the script as constrained domains have a
tendency to cause isValid to fail for the wrong reasons with respect to
the functionality of this script. I.e. isValid should only fail within
this script if the current inner edge is wound incorrectly, failing for
any other reason will cause the resulting domain to be invalid as it
will cause an incorrect edge flip.