Remove useless 'initializeDomain' option
This commit removes the 'initializeDomain' option and the accompanying
logic; it turns out that this did not work the way I thought it did. The
domain will be initialized in any event, regardless of the value
'initializeDomain' is given. I am not sure how this got past me for so
long.